resync from head

+++ sys/arch/arm/cortex/a9tmr_var.h
@@ -1,4 +1,4 @@
-/* $NetBSD: a9tmr_var.h,v 1.1.2.1 2012/11/20 03:01:04 tls Exp $ */
+/* $NetBSD: a9tmr_var.h,v 1.1.2.2 2013/06/23 06:20:00 tls Exp $ */
 /*-
  * Copyright (c) 2012 The NetBSD Foundation, Inc.
  * All rights reserved.
@@ -39,7 +39,7 @@ struct a9tmr_softc {
 	bus_space_handle_t sc_private_memh;
 	bus_space_handle_t sc_wdog_memh;
 	struct evcnt sc_ev_missing_ticks;
-	u_long sc_freq;
+	uint32_t sc_freq;
 	u_long sc_autoinc;
 	void *sc_global_ih;
 };
